Update kit init container entrypoint to support extra flags
Add support for '$EXTRA_FLAGS' environment variable in the Kit init
container to allow specifying arbitrary flags for the unpack command. If
this environment variable is used to specify the --dir or --filter
flags, exit with an error.

By default, use '--ignore-existing' when unpacking to ensure we can run
container with mounted persistent storage without erroring (if files
e.g. exist). This behavior can be disabled by setting an alternate value
for '$EXTRA_FLAGS'

Signed-off-by: Angel Misevski <amisevsk@gmail.com>