Escape braces in CLI documentation when generating markdown
Using {{ term }} in markdown causes issues for vitepress. As a result,
we have to escape these characters when generating cli-reference.md from
the CLI help pages.

Additionally, prefix template directive lines with '*' to make sure they
appear as a list in rendered markdown, even though this looks worse in
the CLI

Signed-off-by: Angel Misevski <amisevsk@gmail.com>