Add dynamic completions for CLI commands
Add completions for locally-stored ModelKits to commands where it makes
sense. This allows completing ModelKit references in e.g. push commands.

The list of ModelKits is generated at run time based on local storage.
For commands that work both locally and remotely, we only generate
completions when the user could be referring to a local ModelKit.

Signed-off-by: Angel Misevski <amisevsk@gmail.com>