Handle unpacking non kitops ModelPacks
If Kit generates a model-spec artifact, it will include the Kitfile with
the manifest. However, if the artifact was generated by another tool, we
can still try to unpack it based on model-spec metadata annotations
(e.g. file names). This is more error-prone, so a warning is logged in
this case.

Signed-off-by: Angel Misevski <amisevsk@gmail.com>