Handle missing tag/digest in ModelKit references per command
Explicitly handle cases where a ModelKit reference is missing a tag or
digest based on usage. This is required as skipping the tag in a
reference is a common error among users, and it helps to clearly report
the issue.

For some commands, we explicitly fail the command, requiring a tag,
with a clear message (e.g. kit remove). For others, we default to using
the tag 'latest', optionally logging an info line to make sure the user
is aware.

Signed-off-by: Angel Misevski <amisevsk@gmail.com>